Refactor parquet extension into separate classes and files and fix glob reading/writing
